api/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0
numpy>=1.24.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Utilities
aiohttp>=3.9.0
httpx>=0.26.0
requests>=2.31.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

